Bart/clean vulkan tutorial (#53)
* Clean up VulkanTutorial code

+++ graphics/src/gris/graphics/vulkan/immediate_context.cpp
@@ -96,7 +96,7 @@ Gris::Graphics::Vulkan::ImmediateContext::operator bool() const
 
 // -------------------------------------------------------------------------------------------------
 
-void Gris::Graphics::Vulkan::ImmediateContext::GenerateMipmaps(const Texture & texture, const vk::Format & imageFormat, int32_t texWidth, int32_t texHeight)
+void Gris::Graphics::Vulkan::ImmediateContext::GenerateMipmaps(const Texture & texture, const vk::Format & imageFormat, uint32_t texWidth, uint32_t texHeight)
 {
     auto const formatProperties = ParentDevice().GetFormatProperties(imageFormat);
 
@@ -123,8 +123,8 @@ void Gris::Graphics::Vulkan::ImmediateContext::GenerateMipmaps(const Texture & t
                                    1))
     };
 
-    auto mipWidth = texWidth;
-    auto mipHeight = texHeight;
+    auto mipWidth = static_cast<int32_t>(texWidth);
+    auto mipHeight = static_cast<int32_t>(texHeight);
 
     for (uint32_t i = 1; i < texture.MipLevels(); i++)
     {
